FingerPrint-main/FingerPrint-main/Code/requirements.txt
altgraph==0.16.1
auto-py-to-exe==2.6.5
bottle==0.12.17
bottle-websocket==0.2.9
certifi==2019.6.16
cffi==1.13.1
chardet==3.0.4
colorama==0.4.1
cx-Freeze==6.0
cycler==0.10.0
Eel==0.10.4
et-xmlfile==1.1.0
future==0.18.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
imutils==0.5.2
kiwisolver==1.1.0
matplotlib==3.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.16
numpy==1.16.3
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
openpyxl==3.1.1
pandas==0.24.2
pefile==2019.4.18
Pillow==6.0.0
protobuf==3.8.0
pycparser==2.19
PyInstaller==3.5
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
pywin32-ctypes==0.2.0
rednose==1.3.0
requests==2.22.0
scipy==1.2.1
six==1.12.0
termstyle==0.1.11
urllib3==1.25.3
whichcraft==0.6.1
XlsxWriter==1.1.8
